Config: validate correct path to system-wide Caffe
A version check for Caffe was added with e2bcb27c; however this version
check was always broken in the caffe_root == SYS case, as the argument
to validate_version() was 'caffe', which is resolved relative to the
current directory.

Change the return value of find_executable() to be the absolute path to
the executable in the found case, and None in the not found case. Then
use this path instead of the hardcoded 'caffe' as the argument to
validate_version(), to allow DIGITS to operate on the Caffe executable
correctly in the SYS case.

Signed-off-by: Luke Yeager <lyeager@nvidia.com>