Change default resize mode to squash
"half_crop" is not used in any deep learning research and should not be
chosen as the default. Also, for some datasets the network can simply
learn the shape of the noise (e.g. one category may be bottles, so the
network learns that all tall and thin images with noise on the left and
right is a bottle)