ConfigFile: do not test file existence of None
Some code paths might result in the ConfigFile module attempting to test
whether None exists as a file, which raises a TypeError. For example, if
DIGITS is run with HOME unset in the environment, as may happen when it
is configured to start automatically in an init script, running with the
uid of a user which has no home directory, the UserConfigFile subclass
initializer will set the filename field to None.

Since it is normal and expected for such use cases to work without any
UserConfigFile, DIGITS should avoid attempting os.path.isfile(None), and
ConfigFile.exists should return False when _filename is None, e.g. as in
the case of a UserConfigFile with HOME unset in the environment.

Signed-off-by: Luke Yeager <lyeager@nvidia.com>