Stop tools from importing global DIGITS install
If you have the digits package installed globally (on your PYTHONPATH)
and you try to run a local DIGITS instance, the tools pick up the global
installation instead of the local one when you start them in a new
process. This patch fixes the issue.

NOTE: in the future, this should be handled by moving the tools scripts
inside the digits package. This is a short-term solution.