Fix some of the open file issues with Torch
When running the image classification model tests like this:

    ./digits-test digits/model/images/classification/test_views.py

the number of open files for the test suite grows to > 900.
You can verify this with:

    while :; do lsof -p $PID | grep tmp | wc -l; sleep 1; done

where PID is the process ID of the test suite.

This patch brings the number of open files down from > 900 to ~90.
There's more work to be done to fix this completely, but this takes a
big step in the right direction.

NOTE: I can't actually run the whole testsuite or I get a bunch of these
OSErrors:

    OSError: [Errno 24] Too many open files