Don't crop image before feeding to DL framework
When using 'crop in form' method, images are cropped in Python
code before 'classify one'. This is slightly inconsistent as
this operation is not performed during training or in the
context of 'classify many'. Besides, this is problematic for
Torch (at least) since the mean image is not cropped and this
makes it difficult to perform mean subtraction during 'classify
one' as the input and mean image have different sizes.

Caffe uses the 'crop_size' option in the data layer transformations.
Similarly, for Torch the crop command-line flag may be used to
crop the image within Torch. Either way, it is not necessary
to crop the image in Python code.