Re-use all weights when re-training on same dataset
Current behaviour:
When a classification model is being re-trained the last `InnerProduct` layer is renamed to allow training on another dataset (with a possibly different number of classes).
This implies that the weights of that last layer are discarded.

Problem:
If the model is being re-trained on the same dataset then initially the model will perform badly as it needs to re-learn the weights of its last `InnerProduct` layer.

Solution:
Refrain from renaming the last `InnerProduct` layer if the model is being re-trained on the same dataset

Note:
This applies to Caffe only.

closes #533