Drop error if a Caffe layer references an invalid bottom
This should help address the case when a layer is added to manipulate e.g. the label blob. In that case the layer must include a directive like:

include {
 phase: TRAIN
}

Otherwise Caffe will drop an error during inference.

A sanity check is added to DIGITS to catch these errors earlier in the process (i.e. before training).