Ignore classification outputs without corresponding label
Since #628 DIGITS does not overwrite the number of outputs in the last fully-connected layer if it is already set to a value.
If the user accidentally specifies too large a `num_output` then inference will fail as reported on #678.
This change causes classification outputs to be ignored if there is no corresponding label.

close #678