Default to Lua table if tds not installed
If tds is not installed because e.g. Torch with installed with torch-nv deb package then default to using a regular Lua table to store LMDB keys.
There is a limit of 2GB for heap memory in Luajit so we might not be able to train on some large datasets without tds.