Add files via upload
Modified the method in which the bounding boxes are passed to the html template such that they are stored in a sorted array rather than an inherently unsorted javascript object. Under the previous version with the unsorted javascript object, the key for the bounding boxes below the image would be incorrectly labeled because the label and color relied on the index in a presumed sorted object.